development-projects/cc-screenshot-ocr/requirements.txt
pillow>=11.3.0
pytesseract>=0.3.13
opencv-python>=4.11.0.86
numpy>=2.3.1
packaging>=25.0